fix schema to uint32

+++ pinecone_datasets/cfg.py
@@ -1,5 +1,4 @@
-from polars.datatypes import Utf8, Int64, Float32, List, Struct, Field
-
+from polars.datatypes import Utf8, Int64, Float32, List, Struct, Field, UInt32
 
 class Storage:
     base_path: str = "gs://pinecone-datasets-dev"
@@ -9,11 +8,11 @@ class Schema:
     documents = {
         "id": Utf8,
         "values": List(Float32),
-        "sparse_values": Struct([Field('indices', List(Int64)), Field('values', List(Float32))])
+        "sparse_values": Struct([Field('indices', List(UInt32)), Field('values', List(Float32))])
     }
     queries = {
         "id": Utf8,
         "values": List(Float32),
-        "sparse_values": Struct([Field('indices', List(Int64)), Field('values', List(Float32))]),
-        "top_k": Int64
+        "sparse_values": Struct([Field('indices', List(UInt32)), Field('values', List(Float32))]),
+        "top_k": UInt32
     }
